[github-actions] use ubuntu-20.04 for github actions (#5871)
This commit update Github Actions workflows to use ubuntu 20.04
whenever applicable.

- Change ubuntu-latest to ubuntu-20.04 to cleanup the annotations:
  Ubuntu-latest workflows will use Ubuntu-20.04 soon. For more
  details, see
  https://github.com/actions/virtual-environments/issues/1816

- Change ubuntu-18.04 to ubuntu-20.04 wherever ubuntu-20.04 works.
  - Fixed a compile issue in tests/unit/test_link_quality.cpp on
    ubuntu-20.04
  - Can not get Simulation/expects work on ubuntu 20.04 due to unknown
    issue. Leave for further enhancement.

+++ tests/unit/test_link_quality.cpp
@@ -333,7 +333,7 @@ void TestRssAveraging(void)
             rss = GetRandomRss();
             IgnoreError(rssAverager.Add(rss));
             sum += rss;
-            mean = static_cast<double>(sum) / j;
+            mean = static_cast<double>(sum) / static_cast<double>(j);
             VerifyOrQuit(ABS(rssAverager.GetAverage() - mean) < 1, "Average does not match the arithmetic mean!");
             VerifyRawRssValue(rssAverager);
             printf("AddRss(%4d) sum=%-5d, mean=%-8.2f RssAverager=", rss, sum, mean);
